+++ libraries/base/src/Data/List/NonEmpty.hs
@@ -94,7 +94,9 @@ module Data.List.NonEmpty (
    , isPrefixOf  -- :: Eq a => [a] -> NonEmpty a -> Bool
    -- * \"Set\" operations
    , nub         -- :: Eq a => NonEmpty a -> NonEmpty a
+   , nubOrd      -- :: Ord a => NonEmpty a -> NonEmpty a
    , nubBy       -- :: (a -> a -> Bool) -> NonEmpty a -> NonEmpty a
+   , nubOrdBy    -- :: (a -> a -> Ordering) -> NonEmpty a -> NonEmpty a
    -- * Indexing streams
    , (!!)        -- :: NonEmpty a -> Int -> a
    -- * Zipping and unzipping streams
@@ -119,6 +121,7 @@ import qualified Prelude
 
 import           Control.Applicative (Applicative (..), Alternative (many))
 import qualified Data.List                        as List
+import qualified Data.List.NubOrdSet              as NubOrdSet
 import qualified Data.Maybe                       as List (mapMaybe)
 import           GHC.Internal.Data.Foldable       hiding (length, toList)
 import qualified GHC.Internal.Data.Foldable       as Foldable
@@ -568,6 +571,13 @@ unzip ((a, b) :| asbs) = (a :| as, b :| bs)
 -- (The name 'nub' means \'essence\'.)
 -- It is a special case of 'nubBy', which allows the programmer to
 -- supply their own inequality test.
+--
+-- This function knows too little about the elements to be efficient.
+-- Its asymptotic complexity is
+-- /O/(/n/ ⋅ /d/), where /d/ is the number of distinct elements in the list.
+--
+-- If there exists @instance Ord a@, it's faster to use 'Data.List.NonEmpty.nubOrd'.
+--
 nub :: Eq a => NonEmpty a -> NonEmpty a
 nub = nubBy (==)
 
@@ -576,6 +586,25 @@ nub = nubBy (==)
 -- function.
 nubBy :: (a -> a -> Bool) -> NonEmpty a -> NonEmpty a
 nubBy eq (a :| as) = a :| List.nubBy eq (List.filter (\b -> not (eq a b)) as)
+
+-- | Same as 'nub', but asymptotically faster, taking only /O/(/n/ log /d/) time.
+-- where /d/ is the number of distinct elements in the list.
+--
+-- @since 4.23.0.0
+nubOrd :: Ord a => NonEmpty a -> NonEmpty a
+nubOrd = nubOrdBy compare
+{-# INLINE nubOrd #-}
+
+-- | Overloaded version of 'Data.List.NonEmpty.nubOrd'.
+--
+-- @since 4.23.0.0
+nubOrdBy :: (a -> a -> Ordering) -> NonEmpty a -> NonEmpty a
+nubOrdBy cmp (y :| ys) = y :| foldr
+  (\x cont seen -> if NubOrdSet.member cmp x seen then cont seen else x : cont (NubOrdSet.insert cmp x seen))
+  (const [])
+  ys
+  (NubOrdSet.insert cmp y NubOrdSet.empty)
+{-# INLINE nubOrdBy #-}
 
 -- | 'transpose' for 'NonEmpty', behaves the same as 'GHC.Internal.Data.List.transpose'
 -- The rows/columns need not be the same length, in which case